Renamed all modules to be prefixed with their language
