+++ .travis.yml
@@ -11,7 +11,7 @@ notifications:
 
 env:
   global:
-    - HE_VERSION=0.21.0
+    - HE_VERSION=1.0.0
 
 branches:
   only:
