+++ lib/modules/ccnumber/index.js
@@ -10,7 +10,7 @@ module.exports = function CCNumber(options) {
   options.patterns = util.defaultValue(options.patterns, () => { return path.join(__dirname, './data'); });
 
   const self = {};
-  self.key = 'ccnumber';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Credit Card Numbers';
   self.description = 'Scans files for potential credit card numbers';
   self.enabled = true;
@@ -35,7 +35,7 @@ module.exports = function CCNumber(options) {
           const nums = rx.toString().split(/[\D]/);
           nums.forEach(function(value){
             result = luhn.validate(value);
-          });    
+          });
         }
         line = 0;
         if(result === true) {
