+++ lib/modules/contents/index.js
@@ -9,7 +9,7 @@ module.exports = function FileContent(options) {
   options.patterns = util.defaultValue(options.patterns, () => { return path.join(__dirname, './data'); });
 
   const self = {};
-  self.key = 'contents';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'File Contents';
   self.description = 'Scans files for dangerous content';
   self.enabled = true;
