+++ lib/modules/entropy/index.js
@@ -9,7 +9,7 @@ module.exports = function Entropy(options) {
   options = util.permittedArgs(options, []);
 
   const self = {};
-  self.key = 'entropy';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Entropy';
   self.description = 'Scans files for strings with high entropy';
   self.enabled = false;
