+++ lib/modules/example-shell/index.js
@@ -7,7 +7,7 @@ module.exports = function Example(options) {
 
   // Set your module definition here
   const self = {};
-  self.key = 'example';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Example Module';
   self.description = 'Example of how to write a module and shell out a command';
   self.enabled = false;
