+++ lib/modules/files/index.js
@@ -8,7 +8,7 @@ module.exports = function Files(options) {
   options = util.permittedArgs(options, ['patterns', 'fileManager']);
 
   const self = {};
-  self.key = 'files';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Secret Files';
   self.description = 'Scans for known secret files';
   self.enabled = true;
