+++ lib/modules/java-find-secbugs/index.js
@@ -8,7 +8,7 @@ module.exports = function FindSecBugs(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'findSecBugs';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'FindSecBugs Scan';
   self.description = 'FindSecBugs find common security issues in Java code.';
   self.enabled = true;
