+++ lib/modules/java-owasp/index.js
@@ -7,7 +7,7 @@ module.exports = function OwaspDependencyCheck(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'dependency-check';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Owasp Dependency Check Scan';
   self.description = 'Scan the dependencies of a Java project.';
   self.enabled = true;
@@ -23,14 +23,14 @@ module.exports = function OwaspDependencyCheck(options) {
       options.logger.warn(`${self.key} scan will not run unless you build the project before`);
       return false;
     }
- 
+
     if(isCompiledJavaProject && !options.exec.commandExists('dependency-check')) {
       options.logger.warn('java files found but dependency-check was not found in $PATH');
       options.logger.warn(`${self.key} scan will not run unless you install Owasp Dependency Check CLI`);
       options.logger.warn('Installation instructions: https://github.com/Stono/hawkeye/blob/master/lib/modules/owaspDependencyCheck/README.md');
       return false;
      }
- 
+
     return isCompiledJavaProject;
   };
 
@@ -50,7 +50,7 @@ module.exports = function OwaspDependencyCheck(options) {
     if(fileManager.exists(mavenTargetFolder)) {
       allFiles = allFiles.concat(fileManager.getAllFilesSync(mavenTargetFolder));
     }
-    
+
     if(fileManager.exists(gradleBuildFolder)) {
       allFiles = allFiles.concat(fileManager.getAllFilesSync(gradleBuildFolder));
     }
@@ -60,7 +60,7 @@ module.exports = function OwaspDependencyCheck(options) {
   const getAbsolutePath = file => {
     return `${fileManager.target}/${file}`;
   };
-  
+
   self.run = function(results, done) {
     const jarFiles = getProjectJars().map(getAbsolutePath).join(' ');
 
@@ -88,7 +88,7 @@ module.exports = function OwaspDependencyCheck(options) {
 
 	    const offenderIndex = dependency.fileName.indexOf(' ');
 	    const offender = dependency.fileName.slice(offenderIndex).trim();
-	    
+
 	    const item = {
 	      code: vulnerability.name,
 	      offender: offender,
@@ -97,10 +97,10 @@ module.exports = function OwaspDependencyCheck(options) {
 	    };
 
 	    results[vulnerability.severity.toLowerCase()](item);
-	    
+
 	  });
 	}
-	
+
       });
 
       done();
@@ -108,4 +108,4 @@ module.exports = function OwaspDependencyCheck(options) {
   };
 
   return Object.freeze(self);
-};    
+};
