+++ lib/modules/node-chs/index.js
@@ -8,7 +8,7 @@ module.exports = function Nsp(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'node-chs';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Node Constant Hashtable Seed check';
   self.description = 'Scans a package.json to check for CHS issues.';
   self.enabled = false;
