+++ lib/modules/node-crossenv/index.js
@@ -8,7 +8,7 @@ module.exports = function Nsp(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'node-crossenv';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Node CrossEnv malware check';
   self.description = 'Scans a package.json for known malicious crossenv packages';
   self.enabled = true;
