+++ lib/modules/node-ncu/index.js
@@ -9,7 +9,7 @@ module.exports = function Ncu(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'ncu';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Node Check Updates';
   self.description = 'Scans a package.json for out of date packages';
   self.enabled = true;
