+++ lib/modules/node-nsp/index.js
@@ -8,7 +8,7 @@ module.exports = function Nsp(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'nsp';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Node Security Project';
   self.description = 'Scans a package.json for known vulnerabilities from NSP';
   self.enabled = true;
