+++ lib/modules/python-bandit/index.js
@@ -7,7 +7,7 @@ module.exports = function Bandit(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'banditScan';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Bandit Scan';
   self.description = 'Bandit find common security issues in Python code.';
   self.enabled = true;
