+++ lib/modules/python-piprot/index.js
@@ -106,7 +106,7 @@ module.exports = function PythonOutdatedDep(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'pythonOutdatedDepScan';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Python Outdated Dependencies Scan';
   self.description = 'Scans a requirements.txt for out of date packages';
   self.enabled = true;
