+++ lib/modules/python-safety/index.js
@@ -7,7 +7,7 @@ module.exports = function Safety(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'safetyScan';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Safety Scan';
   self.description = 'Safety checks your installed dependencies for known security vulnerabilities.';
   self.enabled = true;
