+++ lib/modules/ruby-brakeman/index.js
@@ -7,9 +7,9 @@ module.exports = function Brakeman(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'brakemanScan';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Brakeman Scan';
-  self.description = 'Brakeman statically analyzes Rails application code to find security issues at any stage of development.';
+  self.description = 'Brakeman statically analyzes Rails application code to find security issues.';
   self.enabled = true;
   let fileManager;
