+++ lib/modules/ruby-bundler-scan/index.js
@@ -7,7 +7,7 @@ module.exports = function BundlerScan(options) {
   options.exec = util.defaultValue(options.exec, () => { return new require('../../exec')(); });
 
   const self = {};
-  self.key = 'bundlerScan';
+  self.key = require('path').basename(require('path').dirname(__filename));
   self.name = 'Bundler Scan';
   self.description = 'Scan for Ruby gems with known vulnerabilities';
   self.enabled = true;
